Refactor to run portal.py from root
